Organise component files
