+++ src/components/Expenses/ExpenseDisplay.css
@@ -4,6 +4,4 @@
     margin: 2rem auto;
     width: 50rem;
     max-width: 95%;
-    border-radius: 12px;
-    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
   }
