+++ src/components/Expenses/ExpenseDisplay.js
@@ -1,9 +1,10 @@
 import ExpenseItem from './ExpenseItem';
+import Card from '../UI/Card'
 import './ExpenseDisplay.css';
 
 function ExpenseDisplay(props) {
   return (
-    <div className='expenses'>
+    <Card className='expenses'>
       <ExpenseItem
         title={props.expenses[0].title}
         amount={props.expenses[0].amount}
@@ -25,7 +26,7 @@ function ExpenseDisplay(props) {
         date={props.expenses[3].date}
       />
 
-    </div>
+    </Card>
   );
 }
