+++ src/components/Expenses/ExpenseItem.js
@@ -1,9 +1,10 @@
 import ExpenseDate from './ExpenseDate';
+import Card from '../UI/Card'
 import './ExpenseItem.css';
 
 function ExpenseItem(props) {
   return (
-    <div className="expense-item">
+    <Card className="expense-item">
       <ExpenseDate date={props.date} />
       <div className="expense-item__description">
         <h2>{props.title}</h2>
@@ -12,7 +13,7 @@ function ExpenseItem(props) {
           {props.amount}
         </div>
       </div>
-    </div>
+    </Card>
   );
 }
